+++ crate/utils/src/crypto/wrap/encrypt_decrypt.rs
@@ -5,12 +5,15 @@ use cloudproof::reexport::crypto_core::{
 use cosmian_kmip::kmip::{
     kmip_data_structures::KeyMaterial,
     kmip_objects::Object,
+    kmip_operations::Encrypt,
     kmip_types::{CryptographicAlgorithm, KeyFormatType, RecommendedCurve},
 };
+use tracing::debug;
 
 use crate::{
+    crypto::ecies::EciesEncryption,
     error::{result::CryptoResultHelper, KmipUtilsError},
-    kmip_utils_bail,
+    kmip_utils_bail, EncryptionSystem,
 };
 
 /// Encrypt bytes using the wrapping key
@@ -22,68 +25,108 @@ pub fn encrypt_bytes<R>(
 where
     R: CryptoRngCore,
 {
-    let wrapping_key_block = wrapping_key
-        .key_block()
-        .context("unable to wrap: wrapping key is not a key")?;
-    // wrap the wrapping key if necessary
-    if wrapping_key_block.key_wrapping_data.is_some() {
-        kmip_utils_bail!("unable to wrap keys: wrapping key is wrapped and that is not supported")
-    }
-    let ciphertext = match wrapping_key_block.key_format_type {
-        KeyFormatType::TransparentSymmetricKey => {
-            // wrap using rfc_5649
-            let wrap_secret = wrapping_key_block.key_bytes()?;
-            key_wrap(plaintext, &wrap_secret)
+    debug!(
+        "encrypt_bytes: with object: {:?}",
+        wrapping_key.object_type()
+    );
+    match wrapping_key {
+        Object::Certificate {
+            certificate_value, ..
+        } => {
+            // TODO(ECSE): cert should be verify before anything
+            //verify_certificate(certificate_value, kms, owner, params).await?;
+            debug!("encrypt_bytes: Encryption with certificate: certificate OK");
+            let ecies = EciesEncryption::instantiate_with_certificate("id", certificate_value)?;
+            let request = Encrypt {
+                data: Some(plaintext.to_vec()),
+                ..Encrypt::default()
+            };
+            let encrypt_response = ecies.encrypt(&request)?;
+            let ciphertext = encrypt_response.data.ok_or(KmipUtilsError::Default(
+                "Encrypt response does not contain ciphertext".to_string(),
+            ))?;
+            debug!("encrypt_bytes: succeeded");
+            Ok(ciphertext)
         }
-        KeyFormatType::TransparentECPublicKey => {
-            // wrap using ECIES
-            match wrapping_key_block.cryptographic_algorithm {
-                CryptographicAlgorithm::ECDH => match &wrapping_key_block.key_value.key_material {
-                    KeyMaterial::TransparentECPublicKey {
-                        recommended_curve,
-                        q_string,
-                    } => match recommended_curve {
-                        RecommendedCurve::CURVE25519 => {
-                            let public_key_bytes: [u8; X25519_PUBLIC_KEY_LENGTH] =
-                                q_string.as_slice().try_into().map_err(|_| {
-                                    KmipUtilsError::ConversionError(
-                                        "invalid X25519 public key length".to_string(),
+        Object::PGPKey { key_block, .. }
+        | Object::SecretData { key_block, .. }
+        | Object::SplitKey { key_block, .. }
+        | Object::PrivateKey { key_block }
+        | Object::PublicKey { key_block }
+        | Object::SymmetricKey { key_block } => {
+            // wrap the wrapping key if necessary
+            if key_block.key_wrapping_data.is_some() {
+                kmip_utils_bail!(
+                    "unable to wrap keys: wrapping key is wrapped and that is not supported"
+                )
+            }
+            let ciphertext = match key_block.key_format_type {
+                KeyFormatType::TransparentSymmetricKey => {
+                    // wrap using rfc_5649
+                    let wrap_secret = key_block.key_bytes()?;
+                    key_wrap(plaintext, &wrap_secret)
+                }
+                KeyFormatType::TransparentECPublicKey => {
+                    // wrap using ECIES
+                    match key_block.cryptographic_algorithm {
+                        CryptographicAlgorithm::ECDH => match &key_block.key_value.key_material {
+                            KeyMaterial::TransparentECPublicKey {
+                                recommended_curve,
+                                q_string,
+                            } => match recommended_curve {
+                                RecommendedCurve::CURVE25519 => {
+                                    let public_key_bytes: [u8; X25519_PUBLIC_KEY_LENGTH] =
+                                        q_string.as_slice().try_into().map_err(|_| {
+                                            KmipUtilsError::ConversionError(
+                                                "invalid X25519 public key length".to_string(),
+                                            )
+                                        })?;
+                                    let public_key =
+                                        X25519PublicKey::try_from_bytes(public_key_bytes)?;
+                                    let ciphertext = EciesSalsaSealBox::encrypt(
+                                        rng,
+                                        &public_key,
+                                        plaintext,
+                                        None,
+                                    )?;
+                                    Ok(ciphertext)
+                                }
+                                x => {
+                                    kmip_utils_bail!(
+                                        "Unable to wrap key: wrapping key: recommended curve not \
+                                         supported for wrapping: {x:?}"
                                     )
-                                })?;
-                            let public_key = X25519PublicKey::try_from_bytes(public_key_bytes)?;
-                            let ciphertext =
-                                EciesSalsaSealBox::encrypt(rng, &public_key, plaintext, None)?;
-                            Ok(ciphertext)
-                        }
+                                }
+                            },
+                            x => {
+                                kmip_utils_bail!(
+                                    "Unable to wrap key: wrapping key: key material not supported \
+                                     for wrapping: {x:?}"
+                                )
+                            }
+                        },
                         x => {
                             kmip_utils_bail!(
-                                "Unable to wrap key: wrapping key: recommended curve not \
+                                "Unable to wrap key: wrapping key: cryptographic algorithm not \
                                  supported for wrapping: {x:?}"
                             )
                         }
-                    },
-                    x => {
-                        kmip_utils_bail!(
-                            "Unable to wrap key: wrapping key: key material not supported for \
-                             wrapping: {x:?}"
-                        )
                     }
-                },
+                }
                 x => {
                     kmip_utils_bail!(
-                        "Unable to wrap key: wrapping key: cryptographic algorithm not supported \
-                         for wrapping: {x:?}"
+                        "Unable to wrap key: wrapping key: format not supported for wrapping: \
+                         {x:?}"
                     )
                 }
-            }
+            }?;
+            Ok(ciphertext)
         }
-        x => {
-            kmip_utils_bail!(
-                "Unable to wrap key: wrapping key: format not supported for wrapping: {x:?}"
-            )
-        }
-    }?;
-    Ok(ciphertext)
+        _ => Err(KmipUtilsError::NotSupported(format!(
+            "Wrapping key type not supported: {:?}",
+            wrapping_key.object_type()
+        ))),
+    }
 }
 
 /// Decrypt bytes using the unwrapping key